requirements.dev.txt
<<<<<<< HEAD
# dev dependencies
ipython>=7.0.0,<=8.0.0
ipdb
pudb
pytest-watch

# Code Quality
flake8
pyproject-flake8
black
isort

# docs
mkdocs

# Install project as editable
-e .
=======
ipython
ipdb
pudb
>>>>>>> 199ecad93caddf53d4240b3a889817690898ae9c
